feat: move instant start export to starter download
