+++ src/commands/index.ts
@@ -11,6 +11,7 @@ import pull from "./pull";
 import push from "./push";
 import repo from "./repo";
 import slice from "./slice";
+import starter from "./starter";
 import status from "./status";
 import sync from "./sync";
 import token from "./token";
@@ -63,6 +64,10 @@ export default createCommandRouter({
 		repo: {
 			handler: repo,
 			description: "Manage repositories",
+		},
+		starter: {
+			handler: starter,
+			description: "Download official starters",
 		},
 		type: {
 			handler: type_,
